requirements.txt
pip>=9.0.1
wheel>0.25.0
flask-restplus>=0.9.2
Flask-SQLAlchemy>=2.1
flask-cors>=0.0
pysolr>=3.6.0
networkx>=1.11
matplotlib>=0.0
sparqlwrapper>0.0
gunicorn>=19.0.0
scipy>=0.0
jsonpickle>=0.0
behave>=0.0
jsonpath_rw>=0.0
pytest>=0.0
percache>=0.0
pydotplus>=0.0
cachier>=0.1.26
